+++ app/upgrades.go
@@ -13,6 +13,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/module"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
+	crisistypes "github.com/cosmos/cosmos-sdk/x/crisis/types"
 	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
 	govv1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
 	govv1beta1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1beta1"
@@ -21,6 +22,7 @@ import (
 
 	appparams "github.com/umee-network/umee/v6/app/params"
 	"github.com/umee-network/umee/v6/util"
+	"github.com/umee-network/umee/v6/x/auction"
 	leveragetypes "github.com/umee-network/umee/v6/x/leverage/types"
 )
 
@@ -48,9 +50,41 @@ func (app UmeeApp) RegisterUpgradeHandlers() {
 	app.registerUpgrade6_0(upgradeInfo)
 	app.registerOutdatedPlaceholderUpgrade("v6.1")
 	app.registerOutdatedPlaceholderUpgrade("v6.2")
-	app.registerUpgrade("v6.3", upgradeInfo)
-
+	app.registerUpgrade("v6.3", upgradeInfo, nil, nil, nil)
 	app.registerUpgrade6_4(upgradeInfo)
+
+	app.registerUpgrade6_5(upgradeInfo)
+}
+
+func (app *UmeeApp) registerUpgrade6_5(upgradeInfo upgradetypes.Plan) {
+	planName := "v6.5"
+
+	app.UpgradeKeeper.SetUpgradeHandler(planName,
+		func(ctx sdk.Context, _ upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
+			printPlanName(planName, ctx.Logger())
+
+			// update leverage and metoken params to include burn auction fee share.
+			lparams := app.LeverageKeeper.GetParams(ctx)
+			lparams.RewardsAuctionFee = sdk.MustNewDecFromStr("0.01")
+			if err := app.LeverageKeeper.SetParams(ctx, lparams); err != nil {
+				return nil, err
+			}
+
+			mekeeper := app.MetokenKeeperB.Keeper(&ctx)
+			meparams := mekeeper.GetParams()
+			meparams.RewardsAuctionFeeFactor = 10000 // 100% of fees goes to rewards auction
+			if err := mekeeper.SetParams(meparams); err != nil {
+				return nil, err
+			}
+
+			return app.mm.RunMigrations(ctx, app.configurator, fromVM)
+		},
+	)
+
+	app.storeUpgrade(planName, upgradeInfo, storetypes.StoreUpgrades{
+		Added:   []string{auction.ModuleName},
+		Deleted: []string{crisistypes.ModuleName},
+	})
 }
 
 func (app *UmeeApp) registerUpgrade6_4(upgradeInfo upgradetypes.Plan) {
@@ -175,14 +209,15 @@ func (app *UmeeApp) storeUpgrade(planName string, ui upgradetypes.Plan, stores s
 
 // registerUpgrade sets an upgrade handler which only runs module migrations
 // and adds new storages storages
-func (app *UmeeApp) registerUpgrade(planName string, upgradeInfo upgradetypes.Plan, newStores ...string) {
+func (app *UmeeApp) registerUpgrade(planName string, upgradeInfo upgradetypes.Plan, newStores []string,
+	deletedStores []string, renamedStores []storetypes.StoreRename) {
 	app.UpgradeKeeper.SetUpgradeHandler(planName, onlyModuleMigrations(app, planName))
 
-	if len(newStores) > 0 {
-		app.storeUpgrade(planName, upgradeInfo, storetypes.StoreUpgrades{
-			Added: newStores,
-		})
-	}
+	app.storeUpgrade(planName, upgradeInfo, storetypes.StoreUpgrades{
+		Added:   newStores,
+		Deleted: deletedStores,
+		Renamed: renamedStores,
+	})
 }
 
 // oldUpgradePlan is a noop, placeholder handler required for old (completed) upgrade plans.