+++ x/uibc/client/cli/query.go
@@ -6,6 +6,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/spf13/cobra"
+
 	"github.com/umee-network/umee/v6/util/cli"
 	"github.com/umee-network/umee/v6/x/uibc"
 )